--- assets/drawable/ActionResult/PlaceHolder
@@ -1,1 +1,0 @@
-何かこのフォルダーにファイルを置くときに消してくださいmm
