+++ lib/Common/Save/Data_common_SaveSlot.dart
@@ -1,5 +1,5 @@
 import 'dart:convert';
-import 'SaveManager.dart';
+import 'Data_common_SaveManager.dart';
 
 class SaveSlot {
   late SaveManager _saveManager;
