+++ lib/Home/UI_home_HomeScreen.dart
@@ -6,7 +6,7 @@ import 'package:sosaku/Common/UI_common_GameScreenBase.dart';
 import 'package:sosaku/Home/Provider_home_HomeScreenProvider.dart';
 import 'package:sosaku/NowLoading/UI_nowLoading_NowLoadingScreen.dart';
 import 'package:sosaku/SelectAction/UI_selectAction_SelectActionScreen.dart';
-import 'package:sosaku/Settings/UI_Setting_SettingScreen.dart';
+import 'package:sosaku/Settings/UI_settings_SettingScreen.dart';
 import 'package:sosaku/Title/Controller_title_SlideShowController.dart';
 import 'package:sosaku/Wrapper/Functions_wrapper_TransitionBuilders.dart';
 import 'package:sosaku/l10n/l10n.dart';
