+++ lib/Menu/Controller_menu_MenuController.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 import '../Home/UI_home_HomeScreen.dart';
-import '../Settings/UI_Setting_SettingScreen.dart';
+import '../Settings/UI_settings_SettingScreen.dart';
 import '../Wrapper/Functions_wrapper_TransitionBuilders.dart';
 import 'Provider_menu_MenuScreenProvider.dart';
 import 'UI_Menu_MenuScreen.dart';
